fix: 修复启动定时任务导致无法正常停止应用程序的问题

+++ app/listener/task/ClearChatRequest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace app\listener\task;
 
+use Swoole\Server;
 use Swoole\Timer;
 use app\model\ChatRequest;
 
@@ -15,11 +16,13 @@ class ClearChatRequest
      *
      * @return mixed
      */
-    public function handle($event)
+    public function handle(Server $server)
     {
-        Timer::tick(86400 * 1000, function () {
-            // 清理过期的入群申请（30天）
-            ChatRequest::where('update_time', '<', (time() - 86400 * 30) * 1000)->delete();
-        });
+        if ($server->getWorkerId() === 0) {
+            Timer::tick(86400 * 1000, function () {
+                // 清理过期的入群申请（30天）
+                ChatRequest::where('update_time', '<', (time() - 86400 * 30) * 1000)->delete();
+            });
+        }
     }
 }
